+++ CHANGELOG.md
@@ -34,6 +34,9 @@
 * Change Package Settings display on View Package details record. (UIEH-1159)
 * Add tests for PackageSearchList, ProviderSearchList, TitleSearchList components. (UIEH-1070)
 * Change `loadMoreButton` to `PrevNextButtons`. (UIEH-1163)
+* Add tests for PackageCreate component. (UIEH-1072)
+* Jest+RTL > Test SearchFilters. (UIEH-1075)
+* Jest+RTL > Test CustomLabelField. (UIEH-1076)
 * Fix Root proxy selection. (UIEH-1143)
 
 ## [6.1.1] (https://github.com/folio-org/ui-eholdings/tree/v6.1.1) (2021-07-15)